INS-654: bootstraper refactoring
