+++ configuration/bootstrap.go
@@ -17,7 +17,6 @@
 package configuration
 
 type Bootstrap struct {
-	RootKeys    string
 	RootBalance uint
 }
